+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [#211] ILL request link in sync logs detail modal
+- [#212] Activity charts on status page (sync trends, action distribution)
 
 ## [1.9.0] - 2026-06-11